requirements.txt
Flask
opencv-python-headless
numpy
google-cloud-vision
python-dotenv
